ref: Move a few things into separate files
